Improved speed (slightly)
Rather than uploading the image to google cloud, Base64 string is passed as a parameter in a http POST request to a service which converts data uri into a passable URL on the fly, service is being hosted on heroku by me. Response time is approx 1 second faster, image quality is much higher however.